Moves the "regression" tests to the root level
