+++ test/msw-api/regression/handle-stream.mocks.ts
